+++ packages/thumbico_core/lib/thumbico_core.dart
@@ -7,6 +7,7 @@ library;
 export 'src/read_thumbico.dart' show readThumbico, readThumbicoAsync;
 export 'src/thumbico_exception.dart' show ThumbicoException, ThumbicoFailure;
 export 'src/thumbico_image.dart' show ThumbicoImage;
+export 'src/thumbico_info.dart' show ThumbicoInfo;
 export 'src/thumbico_option.dart' show ThumbicoOption;
 export 'src/thumbico_size.dart' show ThumbicoSize;
 export 'src/thumbico_source.dart' show ThumbicoSource;